Alright moved the set_occupant method, but I need to do cleanup.
