+++ Map_Project/classes/maptile.py
@@ -1,4 +1,4 @@
-from classes.abstract.character import Character
+# from classes.abstract.character import Character
 
 class MapTile():
     """
@@ -20,12 +20,12 @@ class MapTile():
         self.y = int(y)
         self.coordinates = (int(x), int(y))
 
-    def set_occupant(self, occupant: Character):
-        # RE: Type Checking - while I enforce the occupant to be a character,
-        #     I could foresee other "entity" types occupying a map space.
-        if occupant == None:
-            self.occupant = occupant
-        elif not isinstance(occupant, Character):
-            raise TypeError("Cannot make a non-character a map occupant!")
-        else:
-            self.occupant = occupant
+    # def set_occupant(self, occupant: Character):
+    #     # RE: Type Checking - while I enforce the occupant to be a character,
+    #     #     I could foresee other "entity" types occupying a map space.
+    #     if occupant == None:
+    #         self.occupant = occupant
+    #     elif not isinstance(occupant, Character):
+    #         raise TypeError("Cannot make a non-character a map occupant!")
+    #     else:
+    #         self.occupant = occupant
