+++ Map_Project/tests/test_maptile.py
@@ -64,56 +64,56 @@ class TestMapTile(unittest.TestCase):
         with self.assertRaises(ValueError) as context:
             the_best_maptile.set_coordinates(x_cord, y_cord)
 
-    """
-    set_occupant tests
-    """
-    def test_set_occupant_success_case(self):
-        """
-        Basic success case. If this fails we're in a heap of trouble.
-        """
-        heroic_maptile = MapTile()
-        simply_dan = Plebian()
+    # """
+    # set_occupant tests
+    # """
+    # def test_set_occupant_success_case(self):
+    #     """
+    #     Basic success case. If this fails we're in a heap of trouble.
+    #     """
+    #     heroic_maptile = MapTile()
+    #     simply_dan = Plebian()
 
-        # No occupant... YET
-        self.assertEqual(heroic_maptile.occupant, None)
+    #     # No occupant... YET
+    #     self.assertEqual(heroic_maptile.occupant, None)
 
-        heroic_maptile.set_occupant(simply_dan)
+    #     heroic_maptile.set_occupant(simply_dan)
 
-        # Now Dan is an occupant.
-        self.assertEqual(heroic_maptile.occupant, simply_dan)
+    #     # Now Dan is an occupant.
+    #     self.assertEqual(heroic_maptile.occupant, simply_dan)
 
-    def test_set_occupant_n0ne_case(self):
-        """
-        In my infinite wisdom I forgot that we want to be able to set
-        the map_occupant to be "None" - a relevant case for when we
-        remove enities from map.
-        """
-        heroic_maptile = MapTile()
-        simply_dan = Plebian()
+    # def test_set_occupant_n0ne_case(self):
+    #     """
+    #     In my infinite wisdom I forgot that we want to be able to set
+    #     the map_occupant to be "None" - a relevant case for when we
+    #     remove enities from map.
+    #     """
+    #     heroic_maptile = MapTile()
+    #     simply_dan = Plebian()
 
-        # No occupant... YET
-        self.assertEqual(heroic_maptile.occupant, None)
+    #     # No occupant... YET
+    #     self.assertEqual(heroic_maptile.occupant, None)
 
-        heroic_maptile.set_occupant(simply_dan)
+    #     heroic_maptile.set_occupant(simply_dan)
 
-        # Dan is an occupant. All is good.
-        self.assertEqual(heroic_maptile.occupant, simply_dan)
+    #     # Dan is an occupant. All is good.
+    #     self.assertEqual(heroic_maptile.occupant, simply_dan)
 
-        # Now we don't want the occupant to be anything.
-        heroic_maptile.set_occupant(None)
+    #     # Now we don't want the occupant to be anything.
+    #     heroic_maptile.set_occupant(None)
 
-        # There's nothing here now!
-        self.assertEqual(heroic_maptile.occupant, None)
+    #     # There's nothing here now!
+    #     self.assertEqual(heroic_maptile.occupant, None)
 
-    def test_set_occupant_FAILURE_case(self):
-        """
-        We're going to try setting an occupant that isn't a character or is "None".
-        This shouldn't fly, though we may have to make adjustments later.
-        """
-        heroic_maptile = MapTile()
-        epic_string = "I am the ember of eternity!"
+    # def test_set_occupant_FAILURE_case(self):
+    #     """
+    #     We're going to try setting an occupant that isn't a character or is "None".
+    #     This shouldn't fly, though we may have to make adjustments later.
+    #     """
+    #     heroic_maptile = MapTile()
+    #     epic_string = "I am the ember of eternity!"
 
-        with self.assertRaises(TypeError) as context:
-            heroic_maptile.set_occupant(epic_string)
+    #     with self.assertRaises(TypeError) as context:
+    #         heroic_maptile.set_occupant(epic_string)
 
-            self.assertTrue("Cannot make a non-character a map occupant!" in context.exception)
+    #         self.assertTrue("Cannot make a non-character a map occupant!" in context.exception)
